[FIX] 인가 허용 경로 ~/presents > ~/public 변경
